refactor: split 3 large files into focused modules
DatabaseConnection.swift (526 → 327 lines):
- SSHTypes.swift (168) — SSH enums, SSHJumpHost, SSHConfiguration
- SSLConfiguration.swift (40) — SSLMode, SSLConfiguration

ThemeDefinition.swift (765 → 110 lines):
- ThemeColors.swift (414) — 8 color structs (Syntax, Editor, DataGrid, etc.)
- ThemeLayout.swift (252) — 8 layout structs (Fonts, Spacing, Typography, etc.)

AppSettings.swift (476 → 285 lines):
- GeneralSettings.swift (94) — StartupBehavior, AppLanguage, GeneralSettings
- EditorSettings.swift (107) — EditorFont, EditorSettings

No type renames, no API changes, no import changes in consuming files.
